adding the pop feature
now contemplating if it should be remove or pop.
Remove would just remove any index in the array, but then you'd have to know the length of the array to not go beyond, though it won't throw any error if you go beyond, but it won't delete or throw a warning, so what's the essence anyway.

Whereas pop will only remove the last item, just as in any pop method.