finalised, need to make a decision
the decision i need to make is about how the library would be structured, if it is going to use typescript all through and if so how we will validate the inputs with ts help.

Or there would be a javascript dist and a typescript dist for javascript and typescript langs to be used by the developer respectively.